requirements.txt
filelock==3.13.1
fvcore==0.1.5.post20221221
geomloss==0.2.6
hydra-core==1.3.2
hydra-colorlog
joblib==1.3.2
matplotlib==3.7.3
numpy==1.24.4
omegaconf==2.3.0
pandas==2.0.3
pytorch_lightning==2.0.7
PyYAML==6.0.1
PyYAML==6.0.1
ray==2.7.1
scikit_learn==1.3.1
scipy
seaborn==0.13.2
setuptools==68.0.0
thop==0.1.1.post2209072238
torch==2.0.1
torch_ema==0.3
tqdm==4.66.1
tensorboard
tables